--- src/pages/Main/Slide/Slide.scss
@@ -1,68 +1,0 @@
-.main {
-  width: 100%;
-}
-
-.slideWrap {
-  margin: 0 auto;
-  overflow: hidden;
-  width: 100vw;
-}
-
-.slideList {
-  list-style-type: none;
-  white-space: nowrap;
-  font-size: 0;
-  position: relative;
-}
-
-.slideList img {
-  width: 100%;
-  max-width: 100vw;
-  object-fit: cover;
-}
-
-li {
-  display: inline-block;
-  vertical-align: middle;
-  width: 100%;
-  transform: all 0.5s;
-}
-a {
-  display: block;
-  position: relative;
-}
-label {
-  position: absolute;
-  z-index: 10;
-  top: 50%;
-  transform: translateY(-50%);
-  padding: 50px;
-  cursor: pointer;
-}
-
-.left {
-  left: 30px;
-  background-image: url('/public/image/left-arrow.png');
-  background: center center 100% no-repeat;
-  width: 50px;
-  height: 50px;
-  background-size: cover;
-}
-.right {
-  right: 30px;
-  background-image: url('/public/image/right-arrow.png');
-  background: center center 100% no-repeat;
-  width: 50px;
-  height: 50px;
-  background-size: cover;
-}
-
-.buttons {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  /* margin: 0 10px; */
-  top: 70%;
-  right: 50%;
-}
