+++ src/pages/Main/Slide/SlideList.js
@@ -1,0 +1,32 @@
+const SLIDE_LIST = [
+  {
+    id: 1,
+    leftLabel: 'sldie04',
+    img: 'https://github.com/dxxcw/code200-images/blob/main/images/Main/summeron.png?raw=true',
+    alt: 'summerOn',
+    rightLabel: 'slide02',
+  },
+  {
+    id: 2,
+    leftLabel: 'sldie01',
+    img: 'https://github.com/dxxcw/code200-images/blob/main/images/Main/celebrate.png?raw=true',
+    alt: 'celebrate',
+    rightLabel: 'slide03',
+  },
+  {
+    id: 2,
+    leftLabel: 'sldie02',
+    img: 'https://github.com/dxxcw/code200-images/blob/main/images/Main/findyourtaste.png?raw=true',
+    alt: 'findyourtaste',
+    rightLabel: 'slide04',
+  },
+  {
+    id: 3,
+    leftLabel: 'sldie03',
+    img: 'https://github.com/dxxcw/code200-images/blob/main/images/Main/wedeya.png?raw=true',
+    alt: 'wediya',
+    rightLabel: 'slide01',
+  },
+];
+
+export default SLIDE_LIST;
